Merge branch 'master' of https://github.com/carreirap/tcc

# Conflicts:
# fichasordens-domain/src/main/java/br/com/fichasordens/Atendimento.java
# fichasordens-domain/src/main/java/br/com/fichasordens/FichaAtendimento.java

+++ fichasordens-domain/src/main/java/br/com/fichasordens/Atendimento.java
@@ -9,7 +9,7 @@ public class Atendimento {
 	private Date data;
 	private String descricao;
 	private BigDecimal valor;
-	private float duracao;
+	private BigDecimal duracao;
 	private FichaAtendimento fichaAtendimento;
 	
 	public int getSequencia() {
@@ -36,10 +36,11 @@ public class Atendimento {
 	public void setValor(BigDecimal valor) {
 		this.valor = valor;
 	}
-	public float getDuracao() {
+	
+	public BigDecimal getDuracao() {
 		return duracao;
 	}
-	public void setDuracao(float duracao) {
+	public void setDuracao(BigDecimal duracao) {
 		this.duracao = duracao;
 	}
 	public FichaAtendimento getFichaAtendimento() {
